+++ src/main/java/com/example/projecttracker/ProjectTrackerApplication.java
@@ -1,0 +1,13 @@
+package com.example.projecttracker;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ProjectTrackerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ProjectTrackerApplication.class, args);
+	}
+
+}
