+++ src/main/java/com/example/projecttracker/dto/ProjectRequest.java
@@ -1,0 +1,25 @@
+package com.example.projecttracker.dto;
+
+import com.example.projecttracker.model.Project;
+
+public class ProjectRequest {
+private Project project;
+
+public ProjectRequest(Project project) {
+	super();
+	this.project = project;
+}
+
+public Project getProject() {
+	return project;
+}
+
+public void setProject(Project project) {
+	this.project = project;
+}
+
+public ProjectRequest() {
+	
+}
+
+}
