+++ src/main/java/com/example/projecttracker/exception/ResourceNotFoundException.java
@@ -1,0 +1,16 @@
+package com.example.projecttracker.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.NOT_FOUND)
+public class ResourceNotFoundException extends RuntimeException{
+	
+	private static final long serialVersionUID =1L;
+	
+	public ResourceNotFoundException (String message) {
+		
+		super(message);
+	}
+
+}
