+++ src/main/java/com/example/projecttracker/repository/StatusRepository.java
@@ -1,0 +1,11 @@
+package com.example.projecttracker.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.example.projecttracker.model.Status;
+
+@Repository
+public interface StatusRepository extends JpaRepository<Status, Long> {
+
+}
