--- src/main/java/edu/umgc/CaPPMS/controller/HomeController.java
@@ -1,16 +1,0 @@
-package edu.umgc.CaPPMS.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class HomeController 
-{
-
-	@RequestMapping(value="/home")
-	public String home()
-	{
-		return "home";
-	}
-	
-}
