--- src/main/java/edu/umgc/CaPPMS/dto/ProjectAttachmentRequest.java
@@ -1,31 +1,0 @@
-package edu.umgc.CaPPMS.dto;
-
-import edu.umgc.CaPPMS.model.Project;
-
-
-public class ProjectAttachmentRequest {
-	
-	private Project project;
-
-	public Project getProject() {
-		return project;
-	}
-
-
-	public void setProject(Project project) {
-		this.project = project;
-	}
-
-
-	public ProjectAttachmentRequest(Project project) {
-		super();
-		this.project = project;
-	}
-
-	public ProjectAttachmentRequest() {
-		
-	}
-	
-	
-	
-}
