--- src/main/java/edu/umgc/CaPPMS/exception/ResourceNotFoundException.java
@@ -1,16 +1,0 @@
-package edu.umgc.CaPPMS.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class ResourceNotFoundException extends RuntimeException{
-	
-	private static final long serialVersionUID =1L;
-	
-	public ResourceNotFoundException (String message) {
-		
-		super(message);
-	}
-
-}
