+++ src/main/resources/application.properties
@@ -3,7 +3,9 @@ spring.datasource.username=postgres
 spring.datasource.password=password
 spring.jpa.show-sql=true
 
-
+## Hibernate Properties
+# The SQL dialect makes Hibernate generate better SQL for the chosen database
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
 
-spring.jpa.hibernate.ddl-auto=update
+# Hibernate ddl auto (create, create-drop, validate, update)
+spring.jpa.hibernate.ddl-auto = update
