--- src/main/resources/templates/home.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-   <head>
-      <meta charset = "ISO-8859-1" />
-      <link href = "css/styles.css" rel = "stylesheet"/>
-      <title>Spring Boot Application</title>
-   </head>
-   <body>
-      <h4>Welcome to Thymeleaf Spring Boot web application</h4>
-   </body>
-</html>n
