+++ src/test/java/com/example/projecttracker/ProjectTrackerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.projecttracker;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ProjectTrackerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
