--- src/test/java/edu/umgc/CaPPMS/CaPpmsApplicationTests.java
@@ -1,13 +1,0 @@
-package edu.umgc.CaPPMS;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CaPpmsApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
